search_console_webapp/requirements.txt
# requirements.txt - ACTUALIZADO para Python 3.13

# Flask y extensiones
Flask==3.0.3
python-dotenv==1.0.1

# Google APIs (actualizadas)
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.147.0

# Procesamiento de datos (COMPATIBLE con Python 3.13)
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
pytz==2024.2

# SERP API
google-search-results==2.4.2

# HTTP requests
requests==2.32.3

# Logging y utilidades (CORREGIDO - logging es parte de stdlib)
# logging==0.4.9.6  # REMOVIDO - no es necesario

# Para screenshots (actualizado)
playwright==1.48.0

# Seguridad adicional
cryptography==43.0.1

# Para manejo de fechas
python-dateutil==2.9.0

# Cache y sesiones
Flask-Session==0.8.0
redis==5.1.1

# PostgreSQL
psycopg2-binary==2.9.9

# Dependencias adicionales útiles
numpy==2.1.3
urllib3==2.2.3

# Stripe for payments
stripe==12.4.0

# Scheduling
schedule==1.2.2

# ====================================================================
# LLM Providers para Multi-LLM Brand Monitoring
# ====================================================================

# OpenAI (GPT-5)
openai==1.54.4

# Anthropic (Claude Sonnet 4.5)
anthropic==0.39.0

# Google Generative AI (Gemini 2.0 Flash)
google-generativeai==0.8.3

# ====================================================================
# Testing Dependencies
# ====================================================================

# pytest y plugins
pytest==8.3.3
pytest-mock==3.14.0
pytest-cov==6.0.0
pytest-asyncio==0.24.0

# Para mocking y testing
mock==5.1.0